Using safer types

+++ DeforaOS/System/src/libSystem/include/System/string.h
@@ -30,15 +30,15 @@ String * string_new(String const * string);
 void string_delete(String * string);
 
 /* returns */
-int string_length(String const * string);
+size_t string_length(String const * string);
 
 /* useful */
 int string_append(String ** string, String const * append);
-void string_cut(String * string, unsigned int length);
+void string_cut(String * string, size_t length);
 
 int string_compare(String const * string, String const * string2);
 int string_compare_length(String const * string, String const * string2,
-		unsigned int length);
+		size_t length);
 
 String * string_find(String const * string, String const * key);
